+++ packages/math/src/types.ts
@@ -13,7 +13,7 @@ export type Vec3 = [number, number, number];
 export type Vec4 = [number, number, number, number];
 /**
  * A 5-element polygon, representing a quadrilateral with 4 corners (viewport, tile, etc)
- * By convention, the first point should be the same as the last point, for GeoJSON compatibility. */
+ * By convention, the first point must be the same as the last point, for GeoJSON compatibility. */
 export type Quad = [Vec2, Vec2, Vec2, Vec2, Vec2];
 
 /** Bounding box containing minX, minY, maxX, maxY numbers */
@@ -24,7 +24,7 @@ export interface BBox {
   maxY: number;
 };
 
-/** An Object containing `index`, `distance`, and `point` properties*/
+/** An Object containing details about the closest point to a given path */
 export interface Closest {
   /** index of segment along path */
   index: number;
@@ -34,15 +34,21 @@ export interface Closest {
   point: Vec2;
 };
 
-/** Surrounding Rectangle. An Object containing `polygon` and `angle` properties. */
+/** An Object containing details about a surrounding rectangle for a given set of points */
 export interface SurroundingRectangle {
   /** The surrounding rectangle polygon */
   polygon: Quad;
-  /** angle offset from x axis */
+  /** Angle offset from x axis */
   angle: number;
+  /** Centroid of the rectangle (long and short axes intersect here) */
+  centroid: Vec2;
+  /** Short axis of the rectangle (perpendicular to angle, the rect's height direction) */
+  shortAxis: [Vec2, Vec2];
+  /** Long axis of the rectangle (along angle, the rect's width direction) */
+  longAxis: [Vec2, Vec2];
 };
 
-/** Contains essential information about a Tile */
+/** An Object containing details about a Tile */
 export interface Tile {
   /** tile identifier string ex. '0,0,0' */
   id: string;
@@ -56,22 +62,12 @@ export interface Tile {
   isVisible: boolean;
 };
 
-/** An Object containing `index`, `distance`, and `point` properties */
-export interface Edge {
-  /** index of segment along path */
-  index: number;
-  /** distance from point to path */
-  distance: number;
-  /** point along path */
-  point: Vec2;
-};
-
 /** An Object used to return information about the tiles covering a given viewport */
 export interface TileResult {
   tiles: Tile[];
 };
 
-/** Contains the properties that define the transform */
+/** An Object containing details about a Transform */
 export interface TransformProps {
   /** translation in x direction from origin */
   x: number;